# backend/travel-go-backend/services/place/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/checkin/create": {
            "post": {
                "description": "Api lưu checkin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Api lưu checkin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Checkin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
        "/checkin/list": {
            "get": {
                "description": "Api lấy danh sách checkin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Api lấy danh sách checkin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "place_id",
                        "name": "place_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_info": {
                    "$ref": "#/definitions/api_response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "entity.Checkin": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "place": {
                    "$ref": "#/definitions/entity.Place"
                },
                "place_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "entity.Place": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "place_img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PlaceImg"
                    }
                },
                "place_img_checkin": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PlaceImgCheckin"
                    }
                },
                "place_type": {
                    "$ref": "#/definitions/entity.PlaceType"
                },
                "place_type_id": {
                    "type": "string"
                },
                "review": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Review"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.PlaceImg": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "place_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entity.PlaceImgCheckin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "place_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entity.PlaceType": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.Review": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "dislike": {
                    "type": "integer"
                },
                "like": {
                    "type": "integer"
                },
                "place_id": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                },
                "review_img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ReviewImg"
                    }
                }
            }
        },
        "entity.ReviewImg": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}